+++ Assets/Scripts/Application Layer/UnitSystem/Interface/IInventoryForSkill.cs
@@ -1,8 +1,8 @@
 
 public interface IInventoryForSkill 
 {
-    public int GetCurrentCarrot();
+    public long GetCurrentCarrot();
     public long GetCurrentMoney();
-    public void DecreaseCarrot(int _amount);
-    public void DecreaseMoney(int _amount);
+    public void DecreaseCarrot(long _amount);
+    public void DecreaseMoney(long _amount);
 }
