+++ Assets/Scripts/Application Layer/UnitSystem/Interface/IMoneyData.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 public interface IMoneyData
 {
     long money { get; }
-    int carrot { get; }
-    int sunEssence { get; }
-    int moonEssence { get; }
-    int lightningEssence { get; }
+    long carrot { get; }
+    long sunEssence { get; }
+    long moonEssence { get; }
+    long lightningEssence { get; }
 }
